Move some components into their own file
