+++ apps/mobile/src/components/browse/BrowseLocations.tsx
@@ -24,65 +24,6 @@ import Fade from '../layout/Fade';
 import ImportModal from '../modal/ImportModal';
 import { LocationModal } from '../modal/location/LocationModal';
 
-interface BrowseLocationItemProps {
-	location: Location;
-	onPress: () => void;
-	editLocation: () => void;
-}
-
-const BrowseLocationItem: React.FC<BrowseLocationItemProps> = ({
-	location,
-	editLocation,
-	onPress
-}: BrowseLocationItemProps) => {
-	const onlineLocations = useOnlineLocations();
-	const online = onlineLocations.some((l) => arraysEqual(location.pub_id, l));
-	const modalRef = useRef<ModalRef>(null);
-	return (
-		<Pressable onPress={onPress}>
-			<Card style={'h-auto w-[110px] flex-col justify-center gap-3'}>
-				<View style={tw`w-full flex-col justify-between gap-1`}>
-					<View style={tw`flex-row items-center justify-between`}>
-						<View style={tw`relative`}>
-							<FolderIcon size={42} />
-							<View
-								style={twStyle(
-									'z-5 absolute bottom-[6px] right-[2px] h-2 w-2 rounded-full',
-									online ? 'bg-green-500' : 'bg-red-500'
-								)}
-							/>
-						</View>
-						<Pressable onPress={() => modalRef.current?.present()}>
-							<DotsThreeOutlineVertical
-								weight="fill"
-								size={20}
-								color={tw.color('ink-faint')}
-							/>
-						</Pressable>
-					</View>
-					<Text
-						style={tw`w-full max-w-[100px] text-xs font-bold text-white`}
-						numberOfLines={1}
-					>
-						{location.name}
-					</Text>
-				</View>
-				<Text style={tw`text-left text-[13px] font-bold text-ink-dull`} numberOfLines={1}>
-					{`${byteSize(location.size_in_bytes)}`}
-				</Text>
-			</Card>
-			<LocationModal
-				editLocation={() => {
-					editLocation();
-					modalRef.current?.close();
-				}}
-				locationId={location.id}
-				ref={modalRef}
-			/>
-		</Pressable>
-	);
-};
-
 const BrowseLocations = () => {
 	const navigation = useNavigation<
 		BrowseStackScreenProps<'Browse'>['navigation'] &
@@ -145,6 +86,65 @@ const BrowseLocations = () => {
 			</Fade>
 			<ImportModal ref={modalRef} />
 		</View>
+	);
+};
+
+interface BrowseLocationItemProps {
+	location: Location;
+	onPress: () => void;
+	editLocation: () => void;
+}
+
+const BrowseLocationItem: React.FC<BrowseLocationItemProps> = ({
+	location,
+	editLocation,
+	onPress
+}: BrowseLocationItemProps) => {
+	const onlineLocations = useOnlineLocations();
+	const online = onlineLocations.some((l) => arraysEqual(location.pub_id, l));
+	const modalRef = useRef<ModalRef>(null);
+	return (
+		<Pressable onPress={onPress}>
+			<Card style={'h-auto w-[110px] flex-col justify-center gap-3'}>
+				<View style={tw`w-full flex-col justify-between gap-1`}>
+					<View style={tw`flex-row items-center justify-between`}>
+						<View style={tw`relative`}>
+							<FolderIcon size={42} />
+							<View
+								style={twStyle(
+									'z-5 absolute bottom-[6px] right-[2px] h-2 w-2 rounded-full',
+									online ? 'bg-green-500' : 'bg-red-500'
+								)}
+							/>
+						</View>
+						<Pressable onPress={() => modalRef.current?.present()}>
+							<DotsThreeOutlineVertical
+								weight="fill"
+								size={20}
+								color={tw.color('ink-faint')}
+							/>
+						</Pressable>
+					</View>
+					<Text
+						style={tw`w-full max-w-[100px] text-xs font-bold text-white`}
+						numberOfLines={1}
+					>
+						{location.name}
+					</Text>
+				</View>
+				<Text style={tw`text-left text-[13px] font-bold text-ink-dull`} numberOfLines={1}>
+					{`${byteSize(location.size_in_bytes)}`}
+				</Text>
+			</Card>
+			<LocationModal
+				editLocation={() => {
+					editLocation();
+					modalRef.current?.close();
+				}}
+				locationId={location.id}
+				ref={modalRef}
+			/>
+		</Pressable>
 	);
 };
