+++ apps/mobile/src/screens/Locations.tsx
@@ -1,27 +1,15 @@
 import { useNavigation } from '@react-navigation/native';
-import { DotsThreeOutlineVertical, Pen, Plus, Trash } from 'phosphor-react-native';
+import { Plus } from 'phosphor-react-native';
 import { useMemo, useRef } from 'react';
-import { Animated, FlatList, Pressable, Text, View } from 'react-native';
-import { Swipeable } from 'react-native-gesture-handler';
+import { FlatList, Pressable, View } from 'react-native';
 import { useDebounce } from 'use-debounce';
-import {
-	arraysEqual,
-	byteSize,
-	Location,
-	useCache,
-	useLibraryQuery,
-	useNodes,
-	useOnlineLocations
-} from '@sd/client';
-import FolderIcon from '~/components/icons/FolderIcon';
-import Card from '~/components/layout/Card';
+import { useCache, useLibraryQuery, useNodes } from '@sd/client';
 import Empty from '~/components/layout/Empty';
 import Fade from '~/components/layout/Fade';
 import { ModalRef } from '~/components/layout/Modal';
 import ScreenContainer from '~/components/layout/ScreenContainer';
-import DeleteLocationModal from '~/components/modal/confirmModals/DeleteLocationModal';
+import LocationItem from '~/components/locations/LocationItem';
 import ImportModal from '~/components/modal/ImportModal';
-import { LocationModal } from '~/components/modal/location/LocationModal';
 import { tw, twStyle } from '~/lib/tailwind';
 import { BrowseStackScreenProps } from '~/navigation/tabs/BrowseStack';
 import { SettingsStackScreenProps } from '~/navigation/tabs/SettingsStack';
@@ -116,124 +104,5 @@ export const Locations = ({ redirectToLocationSettings }: Props) => {
 			</Fade>
 			<ImportModal ref={modalRef} />
 		</ScreenContainer>
-	);
-};
-
-interface LocationItemProps {
-	location: Location;
-	onPress: () => void;
-	editLocation: () => void;
-	navigation: SettingsStackScreenProps<'LocationSettings'>['navigation'];
-}
-
-export const LocationItem = ({
-	location,
-	editLocation,
-	onPress,
-	navigation
-}: LocationItemProps) => {
-	const onlineLocations = useOnlineLocations();
-	const online = onlineLocations.some((l) => arraysEqual(location.pub_id, l));
-	const modalRef = useRef<ModalRef>(null);
-
-	const renderRightActions = (
-		progress: Animated.AnimatedInterpolation<number>,
-		_: any,
-		swipeable: Swipeable
-	) => {
-		const translate = progress.interpolate({
-			inputRange: [0, 1],
-			outputRange: [100, 0],
-			extrapolate: 'clamp'
-		});
-
-		return (
-			<Animated.View
-				style={[
-					tw`mr-3 flex flex-row items-center gap-2`,
-					{ transform: [{ translateX: translate }] }
-				]}
-			>
-				<Pressable
-					style={tw`items-center justify-center rounded-md border border-mobile-lightborder bg-mobile-button px-3 py-1.5 shadow-sm`}
-					onPress={() => {
-						navigation.navigate('EditLocationSettings', { id: location.id });
-						swipeable.close();
-					}}
-				>
-					<Pen size={18} color="white" />
-				</Pressable>
-				<DeleteLocationModal
-					locationId={location.id}
-					trigger={
-						<View
-							style={tw`items-center justify-center rounded-md border border-mobile-lightborder bg-mobile-button px-3 py-1.5 shadow-sm`}
-						>
-							<Trash size={18} color="white" />
-						</View>
-					}
-				/>
-			</Animated.View>
-		);
-	};
-
-	return (
-		<Pressable onPress={onPress}>
-			<Swipeable
-				containerStyle={tw`rounded-md border border-mobile-cardborder bg-mobile-card`}
-				enableTrackpadTwoFingerGesture
-				renderRightActions={renderRightActions}
-			>
-				<Card style={tw`h-auto flex-row justify-between gap-3 border-0 p-3`}>
-					<View style={tw`w-[50%] flex-row items-center gap-2`}>
-						<View style={tw`relative`}>
-							<FolderIcon size={38} />
-							<View
-								style={twStyle(
-									'z-5 absolute bottom-[6px] right-[2px] h-2 w-2 rounded-full',
-									online ? 'bg-green-500' : 'bg-red-500'
-								)}
-							/>
-						</View>
-						<View>
-							<Text
-								style={tw`w-auto max-w-[160px] text-sm font-bold text-white`}
-								numberOfLines={1}
-							>
-								{location.name}
-							</Text>
-							<Text numberOfLines={1} style={tw`text-xs text-ink-dull`}>
-								{location.path}
-							</Text>
-						</View>
-					</View>
-					<View style={tw`flex-row items-center gap-3`}>
-						<View style={tw`rounded-md bg-mobile-highlight p-1.5`}>
-							<Text
-								style={tw`text-left text-xs font-bold text-ink-dull`}
-								numberOfLines={1}
-							>
-								{`${byteSize(location.size_in_bytes)}`}
-							</Text>
-						</View>
-						<Pressable onPress={() => modalRef.current?.present()}>
-							<DotsThreeOutlineVertical
-								weight="fill"
-								size={20}
-								color={tw.color('ink-dull')}
-							/>
-						</Pressable>
-					</View>
-				</Card>
-			</Swipeable>
-			<LocationModal
-				editLocation={() => {
-					editLocation();
-					modalRef.current?.close();
-				}}
-				locationId={location.id}
-				ref={modalRef}
-			/>
-		</Pressable>
 	);
 };
